+++ web/src/pages/Editor/Toolbar.vue
@@ -130,7 +130,7 @@
         <q-item v-for="exportOption of exportOptions"
                 :key="exportOption.id"
                 clickable
-                @click="showExportDialog(exportOption.id)"
+                @click="showExportDialog(exportOption.id,exportOption.ext)"
                 dense
         >
           <q-item-section>
@@ -157,7 +157,7 @@
         <q-item v-for="importOption of importOptions"
                 :key="importOption.id"
                 clickable
-                @click="importFile(importOption.id)"
+                @click="importFile(importOption.id,importOption.ext)"
                 dense
         >
           <q-item-section>
@@ -232,23 +232,52 @@ onMounted(()=>{
   const exportOptions = ref([
     {
       id: 'json',
-      label: 'Json'
+      label: 'Json',
+      ext:"json"
     },
     {
       id: 'svg',
-      label: 'SVG'
+      label: 'SVG',
+      ext:"svg"
     },
     {
       id: 'png',
-      label: 'PNG'
+      label: 'PNG',
+      ext:"png"
+    },
+    {
+      id: 'dbml',
+      label: 'DBML',
+      ext:"dbml"
+    },
+    {
+      id: 'pg',
+      label: 'Postgres',
+      ext:"sql"
+    },
+    {
+      id: 'mssql',
+      label: 'MSSQL',
+      ext:"sql"
+    },
+    {
+      id: 'mysql',
+      label: 'MySQL',
+      ext:"sql"
     }
   ])
 
   const importOptions = ref([
     {
       id: 'json',
-      label: 'Json'
-    }
+      label: 'Json',
+      ext:'json'
+    },
+    {
+      id: 'pg',
+      label: 'Postgres',
+      ext:'sql'
+    },
   ])
 
   const dark = computed({
@@ -309,10 +338,10 @@ onMounted(()=>{
     })
   }
 
-  const showExportDialog = (id) => {
+  const showExportDialog = (id,ext) => {
     var cfn = files.getCurrentFile.split('/');
-    var fn = cfn[cfn.length-1].split('.')[0] +'.'+ id;
-console.log(cfn);
+    var fn = cfn[cfn.length-1].split('.')[0] +'.'+ ext;
+  console.log(cfn,id,ext);
     $q.dialog({
       component: VDbExportDialog,
       componentProps: {
@@ -322,13 +351,14 @@ console.log(cfn);
     })
   }
 
-  const importFile = (id) => {
+  const importFile = (id,ext) => {
       console.log(id)
 
       $q.dialog({
       component: VDbImportDialog,
       componentProps: {
-        id:id
+        id:id,
+        ext:ext
       },
     })
